requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
charset-normalizer==3.4.2
distro==1.9.0
exceptiongroup==1.2.2
filelock==3.18.0
fsspec==2025.3.2
h11==0.14.0
hf-xet==1.1.0
httpcore==1.0.8
httpx==0.28.1
huggingface-hub==0.31.1
idna==3.10
jinja2==3.1.6
jiter==0.9.0
joblib==1.5.0
markupsafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.5
openai==1.74.0
packaging==25.0
pillow==11.2.1
pydantic==2.11.3
pydantic-core==2.33.1
python-dotenv==1.1.0
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sentence-transformers==4.1.0
setuptools==80.4.0
sniffio==1.3.1
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.51.3
typing-extensions==4.13.2
typing-inspection==0.4.0
urllib3==2.4.0
